Config coverage

+++ spec/spec_helper.rb
@@ -17,6 +17,6 @@ RSpec.configure do |config|
   config.failure_color = :yellow
   config.after(:suite) do
     new_config_path = File.expand_path("../data/new_config.yml", __FILE__)
-    File.delete(new_config_path)
+    File.delete(new_config_path) if File.exist?(new_config_path)
   end
 end
